M/G/1/FIFO/∞/∞ Model v2.4
v2.4: Code optimized to be shorter and faster by using lists instead of arrays. There was also a minor correction in the code.
v2.3: You can now input expressions instead of numbers only. There was also a minor correction in the code.
v2.2: Code optimized to be shorter and faster. (Oldest version available)

This model is also known as M/G/1/FCFS/∞/∞ or M/G/1.
- Arrivals rate fits a Poisson distribution
- Service time fits a general or arbitrary distribution with standard deviation σ (sigma)
- Only one server/channel
- First-In-First-Out or First-Come-First-Served queuing discipline
- Both the system capacity and the total client population are considered infinite